python/requirements.txt
transformers==5.2.0
fastapi==0.131.0
uvicorn==0.41.0
pillow==12.1.1
